Original change: https://googleplex-android-review.googlesource.com/c/platform/packages/modules/Bluetooth/+/21527500

+++ system/gd/os/android/metrics.cc
@@ -443,7 +443,12 @@ void LogMetricBluetoothLocalVersions(
     uint8_t hci_version,
     uint32_t hci_revision) {
   int ret = stats_write(
-      BLUETOOTH_LOCAL_VERSIONS_REPORTED, lmp_manufacturer_name, lmp_version, lmp_subversion, hci_version, hci_revision);
+      BLUETOOTH_LOCAL_VERSIONS_REPORTED,
+      static_cast<int32_t>(lmp_manufacturer_name),
+      static_cast<int32_t>(lmp_version),
+      static_cast<int32_t>(lmp_subversion),
+      static_cast<int32_t>(hci_version),
+      static_cast<int32_t>(hci_revision));
   if (ret < 0) {
     LOG_WARN(
         "Failed for LogMetricBluetoothLocalVersions, "
